egressgw: split Cell definition to cell.go
This follows the general convention of defining a Cell.
